Split tests into their own files too.
